+++ java-generator/core/src/main/java/io/fabric8/java/generator/nodes/JPrimitiveNameAndType.java
@@ -24,10 +24,14 @@ public class JPrimitiveNameAndType extends JavaNameAndType {
   static final JPrimitiveNameAndType FLOAT = new JPrimitiveNameAndType("Float");
   static final JPrimitiveNameAndType DOUBLE = new JPrimitiveNameAndType("Double");
   static final JPrimitiveNameAndType STRING = new JPrimitiveNameAndType("String");
-  static final JPrimitiveNameAndType DATETIME = new JPrimitiveNameAndType(DATETIME_NAME);
+  static final JPrimitiveNameAndType DATETIME_FROM_STRING = new JPrimitiveNameAndType(DATETIME_NAME, "String");
   static final JPrimitiveNameAndType ANY_TYPE = new JPrimitiveNameAndType("io.fabric8.kubernetes.api.model.AnyType");
 
   public JPrimitiveNameAndType(String name) {
     super(name, JavaType.PRIMITIVE);
+  }
+
+  public JPrimitiveNameAndType(String name, String fromCrdType) {
+    super(name, JavaType.PRIMITIVE, fromCrdType);
   }
 }